@@ -4,4 +4,4 @@ contextBridge.exposeInMainWorld("api", {
   close() { ipcRenderer.send("close") },
   maximizeToggle() { ipcRenderer.send("maximizeToggle") },
   minimize() { ipcRenderer.send("minimize") }
-}) 
+})
